Sponsors: /new_sponsor <...> <...>

+++ orienteer/general/data/ss14/services/whitelist.py
@@ -8,8 +8,14 @@ async def check_whitelist(user_id: UUID) -> bool:
 
 
 async def delete_from_whitelist(user_id: UUID) -> str | None:
-    return await whitelist.delete_user(user_id)
+    if not await whitelist.check(user_id):
+        return await whitelist.delete_user(user_id)
+    else:
+        return None
 
 
 async def add_to_whitelist(user_id: UUID) -> str | None:
-    return await whitelist.add_user(user_id)
+    if not await whitelist.check(user_id):
+        return await whitelist.add_user(user_id)
+    else:
+        return None
